Send modals to individual components
